+++ CI/src/integration_tests/test_command_hooks.py
@@ -25,7 +25,7 @@ class TestCommandHooks(unittest.TestCase):
                                         os.environ["CMAKE_INSTALL_PREFIX"] + "/etc/hooks.d/01-dummy-hook.json")):
             hook_output = hook_in_output_of_hooks_command("01-dummy-hook")
             self.assertEqual(hook_output[1], "/opt/sarus/default/bin/dummy_hook")
-            self.assertEqual(hook_output[2], "prestart")
+            self.assertEqual(hook_output[2], "createContainer")
 
     def test_command_hooks_mpi(self):
         expected_header = ["NAME", "MPI", "TYPE"]
@@ -66,7 +66,7 @@ def generate_dummy_hook_config():
         "when": {
             "always": True
         },
-        "stages": ["prestart"]
+        "stages": ["createContainer"]
     }
     return config
 
@@ -84,6 +84,6 @@ def generate_mpi_hook_config(mpi_type):
                 "^com.hooks.mpi.type": f"^{mpi_type}$"
             }
         },
-        "stages": ["prestart"]
+        "stages": ["createContainer"]
     }
     return config