+++ src/hooks/glibc/GlibcHook.cpp
@@ -36,7 +36,6 @@ GlibcHook::GlibcHook() {
     logMessage("Initializing hook", sarus::common::LogLevel::INFO);
 
     std::tie(bundleDir, pidOfContainer) = hooks::common::utility::parseStateOfContainerFromStdin();
-    hooks::common::utility::enterMountNamespaceOfProcess(pidOfContainer);
     parseConfigJSONOfBundle();
     parseEnvironmentVariables();